Added tests for guide curve lofting
The current fusing of guides and profiles in MakePatches is not reliable
due to the buggy OpenCASCADE fusing routine.

In the test case

TEST(guideCurvePatches, makeLoopsSimpleWingSimpleTest)

this leads to a failing test due to non-closed patch boundaries.
However, the test case

TEST_P(guideCurvePatchesMakeLoopsTest, makeLoops)

contains a possible workaround which uses sequential fusing of all
guides and profile edges.