Improved check for closed guides
No fixed number of max loops required anymore. We could add it again, to provide
some safety - i.e. to prevent an infinite loop. The max loops number is
the same as the number of faces created. We can estimate it from the number of
guide and profile edges.

addresses #145