Improved modeling of half fuselages
The new code mirrors the fuselage profile at the x-z plane,
creates the bspline and trims it again at this plane. This way,
we make sure that the connection between both halfs is
continuous and the tangents are horizontal.

Unfortunately, the test case FuselageGuideCurve.tiglFuselageGuideCurve_CCPACSFuselageSegment
is failing and I have no idea why. The test case seems to assume
profiles that have perfect circle shape?!

addresses #125