Reimplemented segment component segment intersection
The new code creates a line in absolute coordinates (instead
of a line in cs coordinates) and computes the "intersection"
with the iso-eta segment line. This is done, by finding the
point on the eta-line, that minimizes the distance to the
cs line.

The new code produces different results! This is intentional!

TODO: Handle cases, were the cs line is outside the chord surface,
i.e. the result xsi not in [0,1].

addresses #157