Fixed an error in CCPACSWingComponentSegment::GetPoint
The error occurred in case of multiple intersections of the eta plane
with the wing. This is the case for e.g. box wings. The current solution
computes all intersection points, projects them to the local y-z plane of
the wing and computes the distance for each point to the eta point.
The point with the smallest distance is returned.

Fixes #176